+++ app/Http/Controllers/Api/Resume/CertificatesController.php
@@ -19,11 +19,12 @@ class CertificatesController extends Controller
     public function index()
     {
         $user =  auth()->user();
-        $profile = $user->profile->first();
+        $profile = $user->profile;
         $profile_id = $profile->id;
         $certificate = Certificate::where("profiles_id", $profile_id)->get();
         $certificateData = $certificate->map(function ($certificate) {
             return [
+                'id' => $certificate-> id,
                 'title' => $certificate->title,
                 'provider' => $certificate->provider,
                 'issueDate' => $certificate->issueDate,
@@ -46,13 +47,13 @@ class CertificatesController extends Controller
      */
     public function store(Request $request)
     {
-        $user = User::where("id", auth()->user()->id)->first();
-        $profile = $user->profile->first();
-        $profiles = $profile->id;
+        $user =  auth()->user();
+        $profile = $user->profile;
+        $profile_id = $profile->id;
 
         $data = [
             'title' => $request->input('title'),
-            'profiles_id' =>$profiles,
+            'profiles_id' =>$profile_id,
             'provider' => $request->input('provider'),
             'issueDate' => $request->input('issueDate'),
             'description' => $request->input('description'),
@@ -93,24 +94,24 @@ class CertificatesController extends Controller
      */
     public function show(Certificate $certificate)
     {
-        $user = User::where("id", auth()->user()->id)->first();
-        $profile = $user->profile->first();
-        if ($certificate->profiles_id == $profile->id) {
-            return response()->json([
-                'success' => true,
-                'message' => 'success',
-                'data' => [
-                    'title' => $certificate->title,
-                    'provider' => $certificate->provider,
-                    'issueDate' => $certificate->issueDate,
-                    'description' => $certificate->description,
-                    'certificateUrl' => $certificate->certificateUrl,
-
-                ],
-            ]);
-        }else{
-
-        }
+//        $user = User::where("id", auth()->user()->id)->first();
+//        $profile = $user->profile->first();
+//        if ($certificate->profiles_id == $profile->id) {
+//            return response()->json([
+//                'success' => true,
+//                'message' => 'success',
+//                'data' => [
+//                    'title' => $certificate->title,
+//                    'provider' => $certificate->provider,
+//                    'issueDate' => $certificate->issueDate,
+//                    'description' => $certificate->description,
+//                    'certificateUrl' => $certificate->certificateUrl,
+//
+//                ],
+//            ]);
+//        }else{
+//
+//        }
 
 
     }
@@ -120,6 +121,16 @@ class CertificatesController extends Controller
      */
     public function update(Request $request, Certificate $certificate)
     {
+        $user = auth()->user();
+        $profile = $user->profile;
+
+        // Kiểm tra xem chứng nhận thuộc về người dùng hiện tại hay không
+        if ($certificate->profiles_id !== $profile->id) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Unauthorized to delete this certificate',
+            ], 403);
+        }
         $data = $request->all();
 
 
@@ -137,10 +148,24 @@ class CertificatesController extends Controller
      */
     public function destroy(Certificate $certificate)
     {
+        $user = auth()->user();
+        $profile = $user->profile;
+
+        // Kiểm tra xem chứng nhận thuộc về người dùng hiện tại hay không
+        if ($certificate->profiles_id !== $profile->id) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Unauthorized to delete this certificate',
+            ], 403);
+        }
+
+        // Xoá chứng nhận
         $certificate->delete();
+
         return response()->json([
             'success' => true,
-            'message' => 'Certificate me deleted successfully',
+            'message' => 'Certificate deleted successfully',
         ]);
     }
+
 }
